TrinityFastAPIDjangoReact/TrinityBackendDjango/requirements.txt
# Django core
Django>=4.2,<5.0
djangorestframework
django-cors-headers

# Multi-tenancy, permissions & history
django-tenants
django-guardian
django-simple-history

# Celery & Redis
celery>=5.2,<6.0
redis

# Databases
psycopg2-binary   # PostgreSQL driver
pymongo           # MongoDB driver

# FastAPI for micro-services
fastapi
uvicorn[standard]

# Arrow Flight
pyarrow>=15.0.0

# Environment
python-dotenv
gunicorn
requests>=2.28.0
minio
pandas
motor
python-multipart
pydantic-settings==2.10.1
httpx
statsmodels
scikit-learn
pykalman
openpyxl
xlsxwriter